Reworked interface and use new TCP listener from SDK
This update syncs the planning project with the SDK. The new tcp listener
is used and the action client to communicate with the client is removed.
The code simplified and more methods from the SDK move interface are used.

The hello world is only rudimentary adapted. A dedicated client library
for the project is actually needed.

Probably the biggest change in behavior is that a job is not aborted
anymore when a move service call fails (except for internal erros).
This allows users to play more with the move service and may result in
more interesting implementations.

The whole package is also reworked using catkin_lint, pylint,
clang-tidy and the CI config is updated.

See the  no_client_action_srv branch for details.