tests: Add a test to run the sample solution in the CI build
Adds the test setup needed to run the path_planner.py script with every
CI build. The added test case simply evaluates the returned job result.

The added test script adds the path_planner.py script to the python path
manually which isn't the cleanest option but cannot be avoid in this
case. The CI Kinetic build job now installs the missing python-typing
dependency which allows for type hints to be added.