iface: several fixes for recent SDK changes
The move command uses our Cartesian path planning, the check_path service
not yet.

The grasp object is placed a little higher above the table, because there
is still an issue where changes to the collision matrix are not saved in the
planning scene.

It shouldn't be necessary to generate intermediate waypoints anymore as this
should be handled by the Cartesian planning method.

The return code of a moveLin() call was not checked.

Signed-off-by: Wolfgang Wiedmeyer <wolfgang.wiedmeyer@kit.edu>