Use `tests_demo_launcher` for CI tests
Update the rll_planning_project to use the new tests_demo_launcher script to launch scripts as tests. The tests are run with and without Gazebo. The tests are now being run with catkin_make to produce more verbose output. The .ci.rosinstall file has been .ci.repos to reflect the change in the file format.